First stab at a self-updating image
This required my changing how the WorkerThread works - instead of
calling a function repeatedly, have it iterate over a generator. This
removes the assumption that the threaded process only ends by being
aborted. We can provide finite (or not) generators now.

The imaging ui is ripe for some organizational improvements too, but
I'm still getting there.