Solve thickening-colorbar-border mystery
Previously, I saw the border around the colorbar getting thicker every
time update_image was called (that is, each time a new scan was made,
or image loaded, etc.) This is because I wasn't clearing the whole
figure, just the axes... and only the image axes, not the colorbar one
too. I needed to do both, or else the old one was still being drawn
underneath.